Have a BaseApp to share the base component - should unify UI tests setup
