+++ app/src/main/java/com/jraska/github/client/GitHubClientApp.kt
@@ -1,31 +1,17 @@
 package com.jraska.github.client
 
-import android.app.Application
-import androidx.lifecycle.ViewModelProvider
 import com.google.firebase.perf.metrics.AddTrace
+import com.jraska.github.client.core.android.BaseApp
 import com.jraska.github.client.core.android.HasServiceModelFactory
 import com.jraska.github.client.core.android.HasViewModelFactory
-import com.jraska.github.client.core.android.ServiceModel
 
-open class GitHubClientApp : Application(), HasViewModelFactory, HasServiceModelFactory {
+open class GitHubClientApp : BaseApp(), HasViewModelFactory, HasServiceModelFactory {
 
-  private val appComponent: AppComponent by lazy { createAppComponent() }
-
-  override fun factory(): ViewModelProvider.Factory {
-    return appComponent.viewModelFactory()
-  }
-
-  override fun serviceModelFactory(): ServiceModel.Factory {
-    return appComponent.serviceModelFactory()
-  }
+  override val appComponent: AppComponent by lazy { createAppComponent() }
 
   @AddTrace(name = "App.onCreate")
   override fun onCreate() {
     super.onCreate()
-
-    appComponent.onAppCreateActions().forEach {
-      it.onCreate(this)
-    }
   }
 
   open fun createAppComponent(): AppComponent {
